Move ParseProtoRequest to util package
It is now also used by the MergeQuerier.
